#32101 change from ->get_record to ->get_records, add safetyflag and check for deleteFiles()

+++ Moosh/Command/Moodle27/File/FileHashDelete.php
@@ -26,30 +26,23 @@ class FileHashDelete extends MooshCommand{
         
         $hash = $this->arguments[0];
         
-        $result = $DB->get_record('files', array('contenthash' => $hash));
+        $result = $DB->get_records('files', array('contenthash' => $hash));
 
-        if ($result === false){
+        if (count($result) == 0){
             cli_error("no file was found");
         }
-        
-        $filesByItemIndex = $DB->get_records('files', array(
-                'itemid'    => $result->itemid,
-                'contextid' => $result->contextid,
-                'component' => $result->component,
-                'filearea'  => $result->filearea,
-            ));
-        
-        if (count($filesByItemIndex) <= 2) {
-            if (count($filesByItemIndex) === 1){
+
+        if (count($result) <= 2) {
+            if (count($result) === 1){
                 // there is just one file, delete it.
-                $this->deleteFiles($filesByItemIndex);
+                $this->deleteFiles($result, true);
             }
             
-            if (count($filesByItemIndex) === 2) {
+            if (count($result) === 2) {
                 // there are 2 files. Check if one of them is '.'
                 $safeDelete = false;
                 
-                foreach ($filesByItemIndex as $file){
+                foreach ($result as $file){
                     if ($file->filename == ".") {
                         $safeDelete = true;
                     }
@@ -58,18 +51,18 @@ class FileHashDelete extends MooshCommand{
                 // If one of them is '.' delete both, otherwise inform user
                 
                 if ($safeDelete) {
-                    $this->deleteFiles($filesByItemIndex);
+                    $this->deleteFiles($result, $safeDelete);
                 } else {
-                    $this->tooManyFiles($filesByItemIndex);
+                    $this->tooManyFiles($result);
                 }
             }
         }else{
             // give user info that there is too many files for safe deletion
-            $this->tooManyFiles($filesByItemIndex);
+            $this->tooManyFiles($result);
         }
     }
 
-    private function deleteFiles(array $files) {
+    private function deleteFiles(array $files, $safeDelete = false) {
         global $DB;
         
         $fileIds = [];
@@ -79,19 +72,18 @@ class FileHashDelete extends MooshCommand{
         }
         
         $where = "id IN (" . implode(',', $fileIds) . ")";
-        
-        $DB->delete_records_select('files', $where);
 
-        echo "Successfully deleted files." . PHP_EOL;
-    }
+        if($safeDelete === true) {
+            $DB->delete_records_select('files', $where);
+            echo "Successfully deleted files." . PHP_EOL;
+            echo "File ID: {$file->id}, contenthash: {$file->contenthash}, itemid: {$file->itemid}, component {$file->component}, filearea {$file->filearea}, filename {$file->filename}" . PHP_EOL;
+        } else {
+            echo "safeDelete is not set. Refuse to remove any records in DB.";
+        }
+}
     
     private function tooManyFiles($files) {
         echo "There are too many files for safe delete.".PHP_EOL;
-        
-        foreach ($files as $file) {
-            echo "File ID: {$file->id}, contenthash: {$file->contenthash}, itemid: {$file->itemid}".PHP_EOL;
-        }
-        
         echo "All files count: " . count($files) . PHP_EOL;
     }
 }
